split out meta feed shuffling into helpers
